refactor: extract recording starts from App
